📦 chore(quality-check.yml): add quality check workflow to automate code quality checks and analysis
This commit adds a new workflow file `.github/workflows/quality-check.yml` to the project. The workflow is triggered on push and pull request events on the `main` branch.

The workflow consists of a single job `quality-check` that runs on `ubuntu-latest` environment. It uses a Docker container `ramageek/image:php8.3-laravel-dev` as the execution environment.

The steps in the job include:
1. Cloning the repository into the `/var/www/html` directory.
2. Verifying the clone operation and displaying the status.
3. Setting up Laravel by copying the `.env.example` file to `.env`, installing dependencies using Composer, generating the application key, and creating a copy of `.env` as `.env.testing`.
4. Running the quality check using the `composer qc` command and configuring SonarQube analysis by copying the `sonar-project.properties.example` file to `sonar-project.properties` and running `sonar-scanner` with the provided login token and host URL.

This workflow aims to automate the process of checking the code quality and performing static code analysis using SonarQube.